First Commit
loadingNewYear is a basic bar that shows the year progress, made in C
under MIT License.

Program now only prints day of year and number of columns of the terminal.
Created main.c, intended to be a clean file, without expressive functions,
bigger functions are stored on source/ files.
terminal.c and .h contain functions to get the terminal size and day of year.